feat(new-create-collective-flow): move sub-components out of section
